--- a/src/utils/parserServer.js
+++ b/src/utils/parserServer.js
@@ -4,7 +4,8 @@ const { existsSync, mkdtempSync } = require("fs");
 const process = require("process");
 const os = require("os");
 
-let SOCKFILE = null;
+let SOCKFILE;
+let NETCAT_ADAPTER;
 
 // Spawn the parser_server.rb subprocess. We do this since booting Ruby is slow,
 // and we can re-use the parser process multiple times since it is statelesss.
@@ -24,8 +25,6 @@ function spawnParserServer(sockfile, env) {
 
   server.unref();
 
-  process.on("beforeExit", (code) => server.kill());
-
   const now = new Date();
 
   // Wait for server to go live
@@ -41,7 +40,7 @@ function generateSockfileName() {
     path.join(os.tmpdir(), `prettier-ruby-${randomId}`)
   );
 
-  return `${tmpdir}/ruby-parser-server.sock`;
+  return `${tmpdir}/prettier-plugin-ruby.sock`;
 }
 
 // Spawns a parser server if it does not exist
@@ -55,12 +54,51 @@ function ensureParseServer(env) {
   spawnParserServer(SOCKFILE, env);
 }
 
+// Checks to see if an executable is available
+function hasCommand(name) {
+  const isWin = require("os").type() === "Windows_NT";
+
+  let result;
+
+  if (isWin) {
+    result = spawnSync("where", [name]);
+  } else {
+    result = spawnSync("command", ["-v", name]);
+  }
+
+  return result.status === 0;
+}
+
+// Finds an netcat-like adapter to use for sending data to a socket. We order
+// these by likelihood of being found so we can avoid some shell-outs.
+function fetchNetcatAdapter() {
+  if (NETCAT_ADAPTER) {
+    return NETCAT_ADAPTER;
+  }
+
+  if (hasCommand("nc")) {
+    NETCAT_ADAPTER = ["nc", "-U"];
+  } else if (hasCommand("telnet")) {
+    NETCAT_ADAPTER = ["telnet", "-u"];
+  } else if (hasCommand("ncat")) {
+    NETCAT_ADAPTER = ["ncat", "-U"];
+  } else if (hasCommand("socat")) {
+    NETCAT_ADAPTER = ["socat", "-"];
+  } else {
+    NETCAT_ADAPTER = ["node", require.resolve("./netcat.js")];
+  }
+
+  return NETCAT_ADAPTER;
+}
+
 // Formats and sends a request for the parser server. We use netcat here since
 // Prettier expects the results of `parse` to be synchrnous, and Node.js does
 // not offer a mechanism for synchronous socket requests. Luckily, netcat is
 // fairly ubuitious at this point.
 function sendRequest(request, env) {
-  const child = spawnSync("nc", ["-U", SOCKFILE], {
+  const [netcatExe, ...netcatArgs] = fetchNetcatAdapter();
+
+  const child = spawnSync(netcatExe, [...netcatArgs, SOCKFILE], {
     env: Object.assign({}, process.env, env),
     input: JSON.stringify(request),
     maxBuffer: 15 * 1024 * 1024 // 15MB