--- a/test/js/ruby/lang.test.js
+++ b/test/js/ruby/lang.test.js
@@ -12,13 +12,16 @@ expect.extend({
 
 test("different lang settings don't break", () => {
   const script = path.join(__dirname, "../../../node_modules/.bin/prettier");
+
+  const env = { ...process.env, LANG: "US-ASCII" };
+  // Make sure a new parser server is spawned
+  delete env.PRETTIER_RUBY_PARSER_HOST;
+
   const child = spawnSync(
     process.execPath,
     [script, "--plugin", ".", "--parser", "ruby"],
     {
-      env: {
-        LANG: "US-ASCII"
-      },
+      env,
       input: "'# あ'"
     }
   );