… test/js/parser.rb