This PR makes the parser process reusable to avoid loading Ruby for each parse invocation. This increases the performance of parsing by ~10x for each file.

## Implementation

In order to reuse the parser, this PR adds a new "parser server" which listens for requests over a unix socket. Whenever we receive a parse request, we first make sure the parser server is running, then proceed the request over the socket rather than spawning a new ruby process.

Currently, this requires that the `netcat` library is installed since Prettier expects `parse` to be sync. Luckily, `nc` is installed  by default on Mac and most Linux distros. I still need to add feature detection for `nc`, so we are still compatible with systems that don't have netcat, etc (:wave: Windows). If anyone knows of a way to support this more robustly, I'm all ears!

## To-Do

- [ ] Setup adapter pattern for sending parse requests so we can support non-netcat systems
- [ ] Add feature detection to enable this
- [ ] Add some tests

## Benchmarks

### Before

```
spec/shopify/coders_spec.rb 210ms
spec/shopify/connector/add_to_collection_spec.rb 211ms
spec/shopify/connector/create_product_spec.rb 212ms
spec/shopify/connector/delete_product_spec.rb 198ms
spec/shopify/connector/export_products_spec.rb 203ms
spec/shopify/connector/get_products_spec.rb 207ms
```

### After

```
spec/shopify/coders_spec.rb 239ms
spec/shopify/connector/add_to_collection_spec.rb 17ms
spec/shopify/connector/create_product_spec.rb 21ms
spec/shopify/connector/delete_product_spec.rb 16ms
spec/shopify/connector/export_products_spec.rb 34ms
spec/shopify/connector/get_products_spec.rb 29ms
```