@kddeisz 2021-01-06
Nice work! Yeah async parser is definitely top of mind.

Could you potentially reuse the async parser I wrote for the test suite? It lives here: https://github.com/prettier/plugin-ruby/blob/master/test/js/parser.rb. That way we don't have 2 async parsers living in the codebase.

netcat is a really inventive solution, nice job!