@kddeisz 2021-01-06
Also probably worth checking out this file https://github.com/prettier/plugin-ruby/blob/119cb026443089eba2c64dc459ef714b1efe2587/bin/server for the PID file usage.